+++ srai/embedders/__init__.py
@@ -2,5 +2,6 @@
 
 from .base import BaseEmbedder
 from .count_embedder import CountEmbedder
+from .gtfs2vec_embedder import GTFS2VecEmbedder
 
-__all__ = ["BaseEmbedder", "CountEmbedder"]
+__all__ = ["BaseEmbedder", "CountEmbedder", "GTFS2VecEmbedder"]